Protected some routes in RoutesProvider.tsx
